+++ src/main/java/org/threadly/concurrent/limiter/package-info.java
@@ -2,5 +2,7 @@
 /**
  * <p>Package which limits thread pools and provides the ability to create sub pools from existing 
  * thread pools.</p>
+ * 
+ * <p>The contents of this package has moved to {@link org.threadly.concurrent.wrapper.limiter}.</p>
  */
 package org.threadly.concurrent.limiter;
