+++ src/main/java/org/threadly/concurrent/wrapper/limiter/AbstractKeyedLimiter.java
@@ -1,4 +1,4 @@
-package org.threadly.concurrent.limiter;
+package org.threadly.concurrent.wrapper.limiter;
 
 import java.util.HashMap;
 import java.util.Map;
