+++ src/main/java/org/threadly/concurrent/wrapper/limiter/package-info.java
@@ -1,0 +1,6 @@
+
+/**
+ * <p>Package which limits thread pools and provides the ability to create sub pools from existing 
+ * thread pools.</p>
+ */
+package org.threadly.concurrent.wrapper.limiter;
