+++ src/test/java/org/threadly/concurrent/KeyDistributedExecutorTest.java
@@ -21,7 +21,7 @@ import org.junit.BeforeClass;
 import org.junit.Test;
 import org.threadly.BlockingTestRunnable;
 import org.threadly.ThreadlyTestUtil;
-import org.threadly.concurrent.limiter.ExecutorLimiter;
+import org.threadly.concurrent.wrapper.limiter.ExecutorLimiter;
 import org.threadly.concurrent.lock.StripedLock;
 import org.threadly.test.concurrent.TestCondition;
 import org.threadly.test.concurrent.TestRunnable;
