+++ src/test/java/org/threadly/concurrent/wrapper/limiter/KeyedSchedulerServiceLimiterTest.java
@@ -1,9 +1,6 @@
-package org.threadly.concurrent.limiter;
+package org.threadly.concurrent.wrapper.limiter;
 
 import static org.junit.Assert.*;
-import static org.threadly.TestConstants.TEST_QTY;
-
-import java.util.Map;
 
 import org.junit.After;
 import org.junit.Before;
@@ -16,8 +13,8 @@ import org.threadly.concurrent.TestCallable;
 import org.threadly.test.concurrent.TestRunnable;
 import org.threadly.util.StringUtils;
 
-@SuppressWarnings({"javadoc", "deprecation"})
-public class KeyedSchedulerServiceLimiterTest {
+@SuppressWarnings("javadoc")
+public class KeyedSchedulerServiceLimiterTest extends AbstractKeyedLimiterTest {
   protected PriorityScheduler scheduler;
   
   @Before
@@ -31,6 +28,7 @@ public class KeyedSchedulerServiceLimiterTest {
     scheduler = null;
   }
 
+  @Override
   protected KeyedSchedulerServiceLimiter makeLimiter(int limit) {
     return new KeyedSchedulerServiceLimiter(scheduler, limit, null, true, 1);
   }
@@ -49,63 +47,6 @@ public class KeyedSchedulerServiceLimiterTest {
       fail("Exception should have thrown");
     } catch (IllegalArgumentException e) {
       // expected
-    }
-  }
-  
-  @Test
-  public void getMaxConcurrencyPerKeyTest() {
-    assertEquals(1, makeLimiter(1).getMaxConcurrencyPerKey());
-    int val = 10;
-    assertEquals(val, makeLimiter(val).getMaxConcurrencyPerKey());
-  }
-  
-  @Test
-  public void getUnsubmittedTaskCountTest() {
-    KeyedSchedulerServiceLimiter singleConcurrencyLimiter = makeLimiter(1);
-    String key = StringUtils.makeRandomString(5);
-    BlockingTestRunnable btr = new BlockingTestRunnable();
-    try {
-      assertEquals(0, singleConcurrencyLimiter.getUnsubmittedTaskCount(key));
-      singleConcurrencyLimiter.execute(key, btr);
-      btr.blockTillStarted();
-      // should not be queued any more
-      assertEquals(0, singleConcurrencyLimiter.getUnsubmittedTaskCount(key));
-      
-      for (int i = 1; i < TEST_QTY; i++) {
-        singleConcurrencyLimiter.submit(key, DoNothingRunnable.instance());
-        assertEquals(i, singleConcurrencyLimiter.getUnsubmittedTaskCount(key));
-      }
-    } finally {
-      btr.unblock();
-    }
-  }
-  
-  @Test (expected = IllegalArgumentException.class)
-  public void getUnsubmittedTaskCountNullFail() {
-    makeLimiter(1).getUnsubmittedTaskCount(null);
-  }
-  
-  @Test
-  public void getUnsubmittedTaskCountMapTest() {
-    KeyedSchedulerServiceLimiter singleConcurrencyLimiter = makeLimiter(1);
-    String key = StringUtils.makeRandomString(5);
-    BlockingTestRunnable btr = new BlockingTestRunnable();
-    try {
-      assertTrue(singleConcurrencyLimiter.getUnsubmittedTaskCountMap().isEmpty());
-      singleConcurrencyLimiter.execute(key, btr);
-      btr.blockTillStarted();
-
-      // should not be queued any more
-      assertTrue(singleConcurrencyLimiter.getUnsubmittedTaskCountMap().isEmpty());
-      
-      for (int i = 1; i < TEST_QTY; i++) {
-        singleConcurrencyLimiter.submit(key, DoNothingRunnable.instance());
-        Map<?, ?> taskCountMap = singleConcurrencyLimiter.getUnsubmittedTaskCountMap();
-        assertEquals(1, taskCountMap.size());
-        assertEquals(i, taskCountMap.get(key));
-      }
-    } finally {
-      btr.unblock();
     }
   }
   
@@ -180,24 +121,6 @@ public class KeyedSchedulerServiceLimiterTest {
   }
   
   @Test
-  public void getCurrentRunningCountTest() {
-    KeyedSchedulerServiceLimiter limiter = makeLimiter(1);
-    String key = StringUtils.makeRandomString(5);
-    
-    assertEquals(0, limiter.getCurrentRunningCount());
-    
-    BlockingTestRunnable btr = new BlockingTestRunnable();
-    try {
-      limiter.execute(key, btr);
-      btr.blockTillStarted();
-      
-      assertEquals(1, limiter.getCurrentRunningCount());
-    } finally {
-      btr.unblock();
-    }
-  }
-  
-  @Test
   public void getQueuedTaskCountTest() {
     // must be single thread scheduler so we can block one on the shceduler
     KeyedSchedulerServiceLimiter limiter = new KeyedSchedulerServiceLimiter(new SingleThreadScheduler(), 1);
@@ -218,32 +141,6 @@ public class KeyedSchedulerServiceLimiterTest {
       limiter.execute(key, DoNothingRunnable.instance());
       // 1 additional blocked in limiter now
       assertEquals(2, limiter.getQueuedTaskCount());
-    } finally {
-      btr.unblock();
-    }
-  }
-  
-  @Test
-  public void getScheduledTaskCountTest() {
-    // must be single thread scheduler so we can block one on the shceduler
-    KeyedSchedulerServiceLimiter limiter = new KeyedSchedulerServiceLimiter(new SingleThreadScheduler(), 1);
-    String key = StringUtils.makeRandomString(5);
-    
-    BlockingTestRunnable btr = new BlockingTestRunnable();
-    try {
-      assertEquals(0, limiter.getScheduledTaskCount());
-      
-      limiter.execute(key, btr);
-      btr.blockTillStarted();
-      
-      limiter.execute(StringUtils.makeRandomString(2), DoNothingRunnable.instance());
-      // 1 blocked on scheduler due to different key
-      assertEquals(1, limiter.getScheduledTaskCount());
-      
-
-      limiter.execute(key, DoNothingRunnable.instance());
-      // 1 additional blocked in limiter now
-      assertEquals(2, limiter.getScheduledTaskCount());
     } finally {
       btr.unblock();
     }
