+++ src/test/java/org/threadly/concurrent/wrapper/limiter/SchedulerServiceLimiterInterfaceTest.java
@@ -1,0 +1,12 @@
+package org.threadly.concurrent.wrapper.limiter;
+
+import org.threadly.concurrent.SchedulerServiceInterfaceTest;
+import org.threadly.concurrent.wrapper.limiter.SchedulerServiceLimiterTest.SchedulerLimiterFactory;
+
+@SuppressWarnings("javadoc")
+public class SchedulerServiceLimiterInterfaceTest extends SchedulerServiceInterfaceTest {
+  @Override
+  protected SchedulerServiceFactory getSchedulerServiceFactory() {
+    return new SchedulerLimiterFactory();
+  }
+}
